+++ sdk/cosmos/azure-cosmos-spark-fabric-account-data-resolver/CHANGELOG.md
@@ -2,7 +2,7 @@
 ### 1.0.0-beta.1 (Unreleased)
 
 #### Features Added
-* Added account data resolver implementation for Azure Cosmos DB Spark Connector. This allows signed in users in Fabric to authenticate to CosmosDB using Microsoft Entra Id. - See [PR 45890](https://github.com/Azure/azure-sdk-for-java/pull/45890)
+* Added account data resolver implementation for Azure Cosmos DB Spark Connector. This allows signed-in users in Fabric to authenticate to CosmosDB using Microsoft Entra ID. - See [PR 45890](https://github.com/Azure/azure-sdk-for-java/pull/45890)
 
 #### Breaking Changes